third_party/blink/web_tests/external/wpt/accname/name/comp_tooltip-expected.txt

This is a testharness.js-based test.
[FAIL] img with tooltip label with empty alt
  assert_equals: <img title="title" data-expectedlabel="title" alt="" data-testname="img with tooltip label with empty alt" class="ex" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw=="> expected "title" but got ""
Harness: the test ran to completion.
